requirements.txt
numpy>=1.20.0
numba>=0.54.0
matplotlib>=3.3.0
quantecon>=0.5.0
scikit-tasmanian>=7.7
pandas>=1.3.0
statsmodels>=0.13.0
